vhost-user: Add generic vhost-user support
The runtime determines that a vhost-user enbaled interface is desired if
the IP for a discovered network interface matches the address appended to a
/tmp/v_ file.

If this port exists and matches, we should take necessary actions
to setup the guest for making use of this vhost-user interface.

In the event that we are attaching to a vhost-user interface, we will need to provide
a mounted hugetablefs for the VM to use.  As an example:
    -object memory-backend-file,id=dpdk-mem,size=1024M,mem-path=/dev/hugepages,share=on,prealloc=on \
    -numa node,memdev=dpdk-mem

On the networking side, an example:
    -chardev socket,id=char1,path=/tmp/v_<ipaddress> \
    -netdev type=vhost-user,id=mynet1,chardev=char1,vhostforce \

The runtime has been tested with OVS and VPP based vhost user interfaces.

Currently it is required to use pc and vmlinuz when using vhost-user,
as issues were identified with our existing clear container setup. More
specifically:
-EFI BIOS does not suppot nvdimm
-pc-lite has multiple bugs and leaks vhost-user memory (appears
that KVM does not release it).
-pc-lite does not shutdown with vhost-user.

Once our machine type supports vhost-user appropriately, we should move
back to using standard clear-containers machine type and fw.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>