networking: Drop pcie bus,slot parameters from qemu interface options
We were assigning a network interface to a particalur pcie bus and
slot to predict its name. Now that we configure based on mac addr,
we can simply drop this allowing qemu to pick the first available pcie
slot. We no longer need to have information about the pcie bus for
a qemu platform in the code.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>