oci: Fix runtime list for pod containers
Pod containers do not have a VM associated to them, but only
a shim.
So we generalize the cc_oci_vm_running() routine to check if
a container is running. For docker containers and pod sandboxes,
this will still check for the VM PID but for pod containers it
will ping the shim PID.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>