runtime: redesign kill command workflow
cc-oci-runtime require changes to be compatible
with docker >=13.1.
kill command will destroy the VM if cc-shim is not running
and the container is not a pod also this command will
change the state to STOPPED if SIGKILL is sent.
If --all option is provided kill command will signal the container
using cc-proxy otherwise the signal will send to cc-shim

Signed-off-by: Julio Montes <julio.montes@intel.com>