sriov: more cleanup
Startup:
 1. Before discovering network, create a HASH of BDF/MAC information to
facilitate identification of Virtual Function PCI devices.

 2. During network creation, for each interface, check to see if it it
is a VF based interface.  We'll know that it is VF based by checking if
the provided interface's MAC address exists in the mac-hash created in
step 1. If so, note the BDF, the driver and mark the net interface
config as being vf-based.  In the event it is VF based, don't do any
further setup.

3. During QEMU command line parameter preperation, for each interface if
it is VF based, unbind it from the host, create a VFIO PCI device and
add appropriate command line specific to using VFIO interface.

4. When creating state file, add a couple of entries into the JSON for
the network device, and more generically for any devices.*

5. During teardown: when shutdown command is received, the stop handler
will bind the device driver back to the host and unbind the given VF
device.  Runtime will then move the VF interface back to the container namespace

Teardown:
When shutdown command is received, the stop handler
will bind the VF PCI device back to the host, and move the VF interface
back to the container namespace.

opens:
1.  What are we passing to the json/statefile? Why do we need both
   cc_oci_devices_to_json and cc_oci_network_devices_to_json; we are
   adding in the same information... looks like it is just overwriting
   what we put there first?

2.  Any further celanup necessary on the tear down path?

Signed-off-by: Eric Ernst <eric.ernst@intel.com>