sriov: fix bind/unbind handling for VFs
Without these changes, when are telling the host
system to bind given device IDs to the vfio-pci driver
by default (when we write the ID to newid file).

After this point, if you recreate the VFs, they'd automatically
bind to VFIO, and not be visible in the network namespace.

To prevent this we now write to the remove_id
for vfio-pci.

Miscellaneous code cleanup / debug messaging added as well.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>