sriov: handle vendor/device ID appropriately, and cleanup
Update / improve sysfs walking, and save the vendor/device id.

This vd_id is now used when adding the newid to the vfio driver,
and for removing the ID during cleanup.

Correctly handle removal of the hash table after it is no longer
needed (ie, post network setup).  may want to consider moving this
to after we fork hypervisor, in order to minimize impact to perceived
boot time?

Signed-off-by: Eric Ernst <eric.ernst@intel.com>